% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmetamod_server.R
\name{Rmetamod_server}
\alias{Rmetamod_server}
\title{Server-side function}
\usage{
Rmetamod_server(input, output, session)
}
\arguments{
\item{input}{see shiny documentation}

\item{output}{see shiny documentation}

\item{session}{see shiny documentation}
}
\description{
Server-side function
}
\seealso{
shiny::shinyApp
}
