% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmetamod_ui.R
\docType{data}
\name{Rmetamod_ui}
\alias{Rmetamod_ui}
\title{UI definition}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 5.
}
\usage{
Rmetamod_ui
}
\description{
UI definition
}
\seealso{
shiny::shinyApp
}
\keyword{datasets}
