% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\docType{data}
\name{UItoSI}
\alias{UItoSI}
\title{Conversion factors from interface (UI) units to SI}
\format{
An object of class \code{numeric} of length 7.
}
\usage{
UItoSI
}
\description{
Conversion factors from interface (UI) units to SI
}
\keyword{datasets}
